Begin Initial Experimentation into Model
Started off with some basic cleaning with a bag of words embedding and
logistic regression classifier, and the results were decent. Saw about a
68% accuracy with about the same recall, but I noticed names of people
and characters were getting emphasized a lot. In the next trial I
replaced all (well, almost all) names in the document with the word
"name" and trained the same model with the same embedding. I saw a
decrease in accuracy to about 66% but with recall still around 68%.
These results also made a lot more sense given the most emphasized words
on the spoiler side were common spoilers such as death and moles,
whereas the important words on the non-spoiler side were about feelings
and common occurrances.